Enhance encrypted file manager

+++ src/main/java/seedu/address/commons/util/EncryptionUtil.java
@@ -1,15 +1,22 @@
 package seedu.address.commons.util;
 
+import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 import java.security.GeneralSecurityException;
+import java.util.Arrays;
 import javax.crypto.Cipher;
 import javax.crypto.SecretKey;
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
 import javax.crypto.spec.PBEParameterSpec;
+
+import seedu.address.commons.exceptions.TargetFileExistException;
 
 /**
  * A class for handling encryption and decryption.
@@ -23,12 +30,13 @@ public class EncryptionUtil {
         DECRYPT
     }
 
-    private static final byte[] salt = {
+    private static final byte[] SALT = {
         (byte) 0x43, (byte) 0x76, (byte) 0x95, (byte) 0xc7,
         (byte) 0x5b, (byte) 0xd7, (byte) 0x45, (byte) 0x17
     };
+    private static final int ITERATION = 68;
 
-    private static final int iteration = 68;
+    private static final String SIGNATURE = "[LOCKED]";
 
     /**
      * Encrypts or decrypts a byte array using a given password.
@@ -42,7 +50,7 @@ public class EncryptionUtil {
         PBEKeySpec keySpec = new PBEKeySpec(password.toCharArray());
         SecretKeyFactory keyFactory = SecretKeyFactory.getInstance("PBEWithMD5AndTripleDES");
         SecretKey key = keyFactory.generateSecret(keySpec);
-        PBEParameterSpec pbeParameterSpec = new PBEParameterSpec(salt, iteration);
+        PBEParameterSpec pbeParameterSpec = new PBEParameterSpec(SALT, ITERATION);
         Cipher cipher = Cipher.getInstance("PBEWithMD5AndTripleDES");
         switch (mode) {
         case ENCRYPT:
@@ -98,34 +106,59 @@ public class EncryptionUtil {
     }
 
     /**
-     * Encrypts a file using the given file path and password.
-     * @param path the path of the file.
+     * Encrypts a file using the given file paths and password.
+     * @param source the path of the original file.
+     * @param target the path of the encrypted file.
      * @param password the password used for decryption.
      * @throws IOException if the encryption fails.
      */
-    public static void encryptFile(String path, String password) throws IOException, GeneralSecurityException {
-        Path oldPath = Paths.get(path);
+    public static void encryptFile(String source, String target, String password)
+            throws IOException, TargetFileExistException, GeneralSecurityException {
+        Path newPath = Paths.get(target);
+        if (Files.exists(newPath)) {
+            throw new TargetFileExistException(target);
+        }
+        Path oldPath = Paths.get(source);
         byte[] fileData = Files.readAllBytes(oldPath);
-        byte[] encryptedData = encryptBytes(fileData, password);
-        Path newPath = Paths.get(oldPath.getParent().toString()
-                + "/[LOCKED] " + oldPath.getFileName().toString());
-        Files.write(newPath, encryptedData);
+        byte[] processedData = encryptBytes(fileData, password);
+        Files.write(newPath, SIGNATURE.getBytes());
+        Files.write(newPath, processedData, StandardOpenOption.APPEND);
         Files.deleteIfExists(oldPath);
     }
 
     /**
-     * Decrypts a file using the given file path and password.
-     * @param path the path of the file.
+     * Decrypts a file using the given file paths and password.
+     * @param source the path of the original file.
+     * @param target the path of the decrypted file.
      * @param password the password used for decryption.
      * @throws IOException if the decryption fails.
      */
-    public static void decryptFile(String path, String password) throws IOException, GeneralSecurityException {
-        Path oldPath = Paths.get(path);
-        byte[] fileData = Files.readAllBytes(oldPath);
-        byte[] decryptedData = decryptBytes(fileData, password);
-        Path newPath = Paths.get(oldPath.getParent().toString()
-                + "/" + oldPath.getFileName().toString().replace("[LOCKED] ", ""));
-        Files.write(newPath, decryptedData);
+    public static void decryptFile(String source, String target, String password)
+            throws IOException, TargetFileExistException, GeneralSecurityException {
+        Path newPath = Paths.get(target);
+        if (Files.exists(newPath)) {
+            throw new TargetFileExistException(target);
+        }
+        Path oldPath = Paths.get(source);
+        byte[] fileData = new byte[(int) Files.size(oldPath) - SIGNATURE.length()];
+        InputStream inStream = new FileInputStream(new File(source));
+        inStream.skip(SIGNATURE.length());
+        inStream.read(fileData, 0, fileData.length);
+        byte[] processedData = decryptBytes(fileData, password);
+        Files.write(newPath, processedData);
         Files.deleteIfExists(oldPath);
+    }
+
+    /**
+     * Returns true if the file is an encrypted file.
+     */
+    public static boolean isFileEncrypted(String path) throws IOException {
+        byte[] fileSignature = new byte[SIGNATURE.length()];
+        InputStream inStream = new FileInputStream(new File(path));
+        int signLength = inStream.read(fileSignature, 0, SIGNATURE.length());
+        if (signLength != SIGNATURE.length()) {
+            return false;
+        }
+        return Arrays.equals(SIGNATURE.getBytes(), fileSignature);
     }
 }
